internal code refactor
