+++ EntityTable/AzureClient/TableEntityBinder.cs
@@ -46,6 +46,7 @@ namespace EntityTableService.AzureClient
 
         public void ReadEntity(ITableEntity entity, IDictionary<string, EntityProperty> properties, OperationContext operationContext)
         {
+
             Entity = new T();
             Metadatas.Clear();
             ReadProp(entity, entity.GetType().GetProperties(), properties);
