+++ EntityTable/EntityTableClient/IEntityTableClient.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EntityTableService.ExpressionFilter;
+using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading;
