+++ EntityTable/ExpressionFilter/Abstractions/IQueryFilter.cs
@@ -1,4 +1,4 @@
-﻿namespace EntityTableService
+﻿namespace EntityTableService.ExpressionFilter.Abstractions
 {
     public interface IQueryFilter<T> : IQueryFilter<T, object> { }
     public interface IQueryFilter<T, P>
