+++ EntityTable/ExpressionFilter/Abstractions/IQueryInstructionsProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace EntityTableService.ExpressionFilter.Abstractions
+{
+    public interface IQueryInstructionsProvider
+    {
+        string Get(string instruction);
+    }
+}
