--- EntityTable/ExpressionHelpers/IQueryInstructionsProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace EntityTableService.ExpressionHelpers
-{
-    public interface IQueryInstructionsProvider : IQueryInstructions
-    {
-        string Get(string instruction);
-    }
-}
